+++ package.json
@@ -9,13 +9,17 @@
     "build": "tsc",
     "prepare": "tsc",
     "pubpreview": "npm pack --dry-run",
+    "publish": "npm publish",
+    "version": "npm version patch",
     "provider": "ts-node provider/SimpleServer.ts",
     "consumer": "ts-node consumer/SimpleClient.ts"
   },
+  "bin-desc": "Provides executable bins for this package.",
   "bin": {
     "web3pay-provider": "provider/SimpleServer.js",
     "web3pay-consumer": "consumer/SimpleClient.js"
   },
+  "fildes-desc": "Whitelist of files which will be published to NPM repository.",
   "files": [
     "lib/**/*",
     "provider/**/*",
@@ -29,6 +33,7 @@
   },
   "devDependencies": {
     "@types/node": "^18.0.6",
+    "pkg": "^5.8.0",
     "ts-node": "^10.9.1",
     "typescript": "^4.7.4"
   }
